refactor(@embark-utils): move logHandler and normalizeInput in utils
