--- packages/embark-utils/src/escapeHtml.js
@@ -1,12 +1,0 @@
-function escapeHtml(message) {
-  if(typeof message !== "string") return message;
-
-  return message
-    .replace(/&/g, "&amp;")
-    .replace(/</g, "&lt;")
-    .replace(/>/g, "&gt;")
-    .replace(/\"/g, "&quot;")
-    .replace(/\'/g, "&#39;");
-}
-
-module.exports = escapeHtml;
