+++ packages/embark/src/lib/modules/process_logs_api/index.js
@@ -1,5 +1,4 @@
-const LogHandler = require('../../utils/logHandler');
-const {escapeHtml} = require('embark-utils');
+const {escapeHtml, LogHandler} = require('embark-utils');
 
 class ProcessLogsApi {
   constructor({embark, processName, silent}) {
